conda-requirements.txt
git
pygithub
requests
tini
tornado
